semver support at runtime for eyeglass modules This change adds semver support to eyeglass modules, avoiding the problems typically associated with peerDependencies. Instead, an eyeglass module can add to their package.json:
```
{
  "eyeglass": {
    "engine": "SEMVER_VALUE"
  }
}
```

And set `SEMVER_VALUE` to a semver compatible string. When eyeglass wraps the
sass options and eyeglass modules are loaded, the `eyeglass.engine` values are
checked against the currently running eyeglass version. If an incompatible
engine is found, the user has the option to stop the node-sass process from
continuing.

To force a hard stop, set `eyeglassStrictVersions: true` in your node-sass
options.

To silence all messages, set `eyeglassStrictVersions: false` in your node-sass
options.

By default, warnings are thrown instead of halting, and the value of
`eyeglassStrictVersions` is `"warn"`.