Cleans up files.js to reduce duplicates and improve readability
The original implementation of files.js that fixed #37 was not as maintainable.
This code change improves the readability of the files.js file by splitting
the namify calls out into their individual sections.

if-return checks were added to the individual namify lines to remove duplication
of file names to improve performance.