Deprecate multiple input paths.
Users should now use broccoli-merge-trees (or broccoli-funnel or both)
to create a single tree for compilation. Users passing multiple trees
will get a deprecation warning and have their trees passed to
broccoli-merge-trees with the overwrite option set. In rare cases, this
can cause a (breaking) change to stylesheet output when several trees
have a file of the same relative name. This can be worked around by
constructing a merged tree with namespaced subdirectories. If files
need to be imported but not directly compiled, it is *far* better to use
the includePaths option instead of passing several trees.