gh-pages now builds successfully
This addresses one half of #183. We needed to both (1) be able to
generate the gh-pages branch again and (2) be able to map eyeglass.rocks
to the new `A` record from github. When attempting a rebuild with node
v8 and v10, the following problems were discovered:

* `gulp-gh-pages` uses an outdated version of `gift`
* `eyeglass-dev-site-builder` needs updated `eyeglass` & `node-sass`

This was fixed by changing the dependency on `gulp-gh-pages` to the
community module `gulp-gh-pages-will`. To fix the eyeglass dependency,
the values in `package-lock.json` were manually increased to a version
of `eyeglass` and `node-sass` that do not run into an Abort Trap on OSX.