[eyeglass] Add assertEyeglassCompatibility option.
This adds support for an option that asserts eyeglass addon
compatibility and defaults to 2.0 being compatible with 1.x. This option
is only used if it would make a newer version of eyeglass compatible
compared to what's in the needs field. This protects against the future
use case where eyeglass 2.x is old and 3.x is already released and some
addons require it.