Introduce a process-wide cache for results from fs.realpath.
In large builds there's often files that are processed many times.
There's enough checks for the realpath of the same file to make caching
worth it. This cuts between 4-5 seconds off of the LinkedIn build.

Results are cached for up to 5 minutes but memory used by the cache is
capped to 1MB.