Additional outputs outside the broccoli tree.
If additional output from a sass compilation is written to a location
that is outside of the broccoli output tree we can't just ignore it
because the persistent cache hit would not create the same outputs. But
because the output locatio is outside of the tree, it's not this
plugin's responsbility to restore that output.

To manage this, the "additional-output" event can now provide a source
location and target url for the asset. When there's a persistent cache
hit, the "cached-asset" event is fired with the source location and
target url that was provided by "additional-output". In this way the
build integration that is using this plugin must decide how to restore
the cache for those files correctly.