Optimize eyeglass asset installation.
Eyeglass assets are installed for each invocation of asset-url() which
can add performance overhead at scale especially when the same asset is
installed many times.

ember-cli-eyeglass now intercepts asset installation and records which
assets are installed to which locations. It does this by writing all
those assets into a single tree for the entire app and all addons and
engines.

This has the benefit of also ensuring that asset urls can be
resolved to any path instead of forcing those urls to resolve to a
location relative to the current tree, which fixes a long-standing
annoyance that forced assets to be duplicated in engines (note: this
patch doesn't change the asset installation locations -- a custom
resolver is required to accomplish this for now).