[ember-cli-eyeglass] per addon and app storage refactor.
In a multi-app project layout, the postBuild hook gets a project that
isn't used elsewhere which makes clearing the app caches more tricky.

instead we use a map to store the information and some global data for
bookkeeping so we can find our data later.

Use the app storage to pass a session cache to broccoli-eyeglass and
reset it after each build.