Be smarter about cache keys for inDevelopment eyeglass addons.
We were using hash-for-dep to invalidate the cache if any javascript
in an eyeglass module changed. But many eyeglass addons are hybrid
packages that have a lot of javascript that isn't relevant to the sass
files.

This patch switches to creating a list of dependencies of js files
starting with the eyeglass exports file and only considers those files
as relevant to the sass file cache. This approach is slower because it
parses the js files for import and require statements, but because
it is a more narrow cache, that extra cost pays for itself.

This only affects eyeglass modules marked as inDevelopment which
use an eyeglass exports file.