Move ember test app & addons to private workspaces
The ember-cli-eyeglass addon's package.json file was serving dual
purposes. It was a dummy ember app for testing the addon as well as the
addon itself. The ember-cli requirements for how to structure the dummy
application required listing some in-repo addons as dependencies of the
dummy application that weren't runtime dependencies for the addon. This
caused weird warnings to be issued by ember-cli.

To make this work better, I've made a proper ember application in
private-packages/eyeglass-test-app and I've moved the in-repo engines to
be private-packages as well. Because we use yarn workspaces, listing
these dependencies as expected in package.json causes everything to just
work as expected. This allows us to make the ember-cli-eyeglass addon
just a simple addon with no ember application embedded within it.

Thanks to @runspired for the advice to use this approach.