fix: Don't add manual modules if they already exist.
BREAKING CHANGE: this fix might change the resolved version of an
eyeglass module that is both a direct dependency, a transitive dependency,
and/or for which a manual module of the same name is provided.

Manual module declarations are added as direct dependencies to
the root. But if the manual module was already in the module tree as
a transitive dependency and a direct dependency at different versions,
then adding the manual module of the transitive dependency would
overwrite the direct dependency. It was never the intent for manual
modules to be used to replace discovered modules, so this behavior
is considered a bug which has now been fixed.

This change has three effects:

1. Manual modules that are path references to a module discovered in the
   node module tree are no longer added as direct dependencies of the
   root module.
2. Manual modules that have the same name as a direct dependency of the
   root module from the package.json dependencies are ignored.
3. If multiple manual modules are added which have the same name, the
   first module of the same name in that list will be added to the root
   module. (Previously, the last module in the list would win.)