Enable esModuleInterop for all packages.
This allows CommonJS to interop with module imports via babel.

It caused some issues in some deprecated code that I was about to remove
anyways, so it's removed as part of this commit:

BREAKING CHANGE: The following deprecated properties are removed:

* require('eyeglass').Eyeglass
* require('eyeglass').decorate

The tests that relied on those deprecations to test our deprecation
warnings have been rewritten to be more generic.