fix: Use prev. asset bundling code if Classic.
7.0.3 introduced changes to the asset bundling code to support Embroider
builds. At the time bundling the assets was conditional based on whether
Classic or Embroider was being used. This changed in 7.0.6, since in
Embroider there was a timing issue with assets sometimes not showing up
in the final artifact.

However, this appears to have caused another issue where extra assets
that shouldn't be present are showing up in the output tree. To fix
this, this commit re-introduces the asset code that was removed in 7.0.3
and conditionally picks what asset build pipeline to use based on
whether the app is being built using Classic or Embroider.